Author: David M <captjay@superlink.net> Log message: - added configuration option to allow opers to set modes on local channels even if they are not oped.
- added configuration option to allow opers to walk through modes on local
channels, since they are required to be on a channel to be able to set
modes on it.

- added configuration option to prevent opers from being deoped from local
channels.

- added configuration option to allow opers to bypass the maximum number of
channels per user.

Detailed behavior changes:

If the admin so choses, opers can now use /mode on local channels even if
they are not oped. In that case, a server notice is generated (SNO_HACK4),
and to clients, it appears that the local server set the mode. The server
notice is not propagated to other servers.

If activated in the configuration, opers can join local channels under any
circumstances. Doing so does not generate a server notice. Note that the
oper can still not talk in the channel if he is banned.

Preventing opers from being deoped on local channels results in the user
attempting the deop to receive a new numeric reply (498) stating he cannot
kick or deop an oper on a local channel. An oper can deop himself, though.
Provided again to fill the service gaps on local channels, allowing admins
who want it to implement "services" for their local channels.


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/trunk@10 c9e4aea6-c8fd-4c43-8297-357d70d61c8c