Author: BLMet <dirk@staf.planetinternet.be> Log message:
Ok after upgrading to ircu 2.10.10.06 yesterday one of my opers today told
me he could not connect anymore today while everything was fine yesterday.
The server dropped him with "Too many connections".

After checking I saw he was in a class that allowed 9 users max from the
same IP. There were 2 connections globally from that IP and he could not
get on. I saw Isomer on #coder-com and he said it had been seen before but
he did not know what it was. Well I found it.

When somebody connects and get throttles there is a ++entry->connected
done in ip_registry_check_local but when you are dropped for throttling
there is no --entry->connected. So when an IP gets throttled more than the
amount of allowed connections for that I-line that IP can not connect
anymore.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_10_pl7@165 c9e4aea6-c8fd-4c43-8297-357d70d61c8c