Author: Kev <klmitch@mit.edu> Log message:
GLINE storm fix...yeah, this is pretty complicated :/  Basically, if
there's a gline for a user we send a NICK for, we make sure and insert
%<lastmod>:<mask> after the user modes, if present.

Status: compiled, tested

Testing needed:

Test all the code paths thoroughly; this is a rather complicated change, so
take nothing for granted.  In particular, make sure bursting servers send
the right stuff, especially in the no-usermode case, which is the only part
I haven't stringently tested.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@248 c9e4aea6-c8fd-4c43-8297-357d70d61c8c