Author: Isomer <isomer@coders.net> Log message:
Well, heres my IPcheck changes.
First note that this INTENTIONALLY doesn't compile - to people that like to
regularly have the bleeding edge version to bug test - thanks, but this
version is just far too unstable, I don't like committing knowingly buggy
code, but this is it. (On the other hand, changing a #if 1 to #if 0 makes it
compile again if your really feeling insane).

I'm interested in various coders to have a look at this, currently this
doesn't deal with nick collisions at all, is this my bug? or something left
over from what kev did?

Theres assert()'s all over the place, if it miscounts then you should have
a nice core for you to examine.


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/trunk@261 c9e4aea6-c8fd-4c43-8297-357d70d61c8c