Author: Kev <klmitch@mit.edu> Log message:
do_names() was adding a space before checking to see if the user was a
zombie; on large channels and in the face of mass kicks, this could cause
the buffer to be overflowed, resulting in a crash of the server.  This
fixes it.

Testing done: it compiles with no warnings, and the server runs and
responds to /names properly; I do not have the resources to test to see if
the problem is completely fixed by this, but it seems apparent that it does


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_10_pl11@278 c9e4aea6-c8fd-4c43-8297-357d70d61c8c