Author: Kev <klmitch@mit.edu> Log message:
Fixed a couple of bugs in the new mode parser, in particular a bug in
mode_parse_limit that wiped out all modes when you did -l, and another
more minor bug having to do with presentation--i.e., if you did a +p
on a channel that was not +s, you would see "-s+p"

Some behavior may still be unexpected; for instance, "mode #test +ps" will
toggle the channel mode between +p and +s, because of the mutual
exclusivity, and if you do "mode #test -n+n" on a +n channel, you will get
"mode #test -n".  We may be able to solve this by removing the shortcut
test at the top of mode_parse_mode...

Testing done: I checked to see if the -iklmntp * behavior worked correctly,
and it did--note that "*" there will be properly treated as an error if the
channel key is, say, "foobar" ;)  I also checked to see that +p/+s worked
as I desired, and it does--which is how I know of the other behavior ;)

Testing needed: tell me what kind of behavior your prefer!


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/trunk@298 c9e4aea6-c8fd-4c43-8297-357d70d61c8c