Author: Kev <klmitch@mit.edu> Log message:
Further rewrites of the MOTD stuff--now the API is fairly clean and well
compartmentalized, so it should be able to work into even a completely new
.conf structure without much trouble at all.  ircd.motd and remote.motd are
read in on initialization and rehash, but T-lines aren't cached unless they
get used, to try to save memory.  There's a new motd_recache() routine that
flushes just the cached MOTDs; this is currently not hooked into the rehash
handler.  Also, we're using ISO dates.

Testing done: Compile-checked; I have not had time to do more, yet.
Status: unstable
Testing needed: Compile it and make sure it doesn't crash when you log in,
  either with or without NODEFAULTMOTD.  Check that T-lines with both
  hostmasks and class designations work.  Check that /rehash or kill -HUP
  don't crash the server.  Watch memory usage to see if there's any
  indications of memory leaks (hard).  Verify that caches get cleared on
  /rehash.  Verify that dates and times are correct and correctly
  displayed.  Verify that /stats T reports all your T-lines and in the
  correct order, with a mix of class- and hostmask-based T-lines.  Check
  what happens if your motd is 0 length.  Basically, try to exercise every
  code path in motd.c and the bits of code that I changed :)


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@314 c9e4aea6-c8fd-4c43-8297-357d70d61c8c