Author: Kev <klmitch@mit.edu> Log message:
Started working on a new logging API that can log to files and/or to
syslog; the intent will be to eventually make this flexible enough that
log files for subsystems can be set from the .conf.

Status: Stable
Testing done:

Compiles.  No run-time checking done because nothing calls these functions
yet.

Future directions:

An API for setting log files needs to be added.  It would also be nice to
have support for sending the log messages through the server notices
mechanism.


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/trunk@316 c9e4aea6-c8fd-4c43-8297-357d70d61c8c