Author: Kev <klmitch@mit.edu> Log message:
Finishing up work on the logging API; lots of dead code added to ircd_log.c
which is intended to be called from appropriate places in the configuration
file parser (I'll add that later).  Started switching bits and pieces over
to the new API.

Status: mostly stable

Testing done:

Compiles and seems to work; some of the functionality can't be tested until
some more work gets done.  The eventual upshot of all this work will be the
eventual replacement of GPATH and others with configuration file options.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@317 c9e4aea6-c8fd-4c43-8297-357d70d61c8c