Author: Kev <klmitch@mit.edu> Log message:
Finished most of the logging system (except for conversions to log_write())
by implementing the feature system, composed of F-lines, to set features
from the .conf, and the three new commands /set, /reset, and /get, to
manipulate feature settings on the fly.

Status: Stable, but expect bugs

Testing needed:

I'm too tired to try to figure out what sort of bugs might be hanging
around.  Beat on it for a while...


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@318 c9e4aea6-c8fd-4c43-8297-357d70d61c8c