Author: Kev <klmitch@mit.edu> Log message:
Implement a msgq system and switch to it, removing some old dead code in
the process--this includes removing help from Configure.help that no longer
has a corresponding compile-time option, as well as permanently switching
to the new mode implementation.

One of the effects this has is a switch to a priority-queue system, which
should allow wallops to be moved to the head of the sendq.  Lag no longer
becomes an issue for important routing notifications.

Status: stable

Testing needed:

This patch needs to be tested under heavy load.

Testing done:

Compiles and runs with no apparent problems


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@325 c9e4aea6-c8fd-4c43-8297-357d70d61c8c