Author: Kev <klmitch@mit.edu> Log message:
I'm fed up with struct Client as it is.  I'm going to restructure the
interface quite a bit, though the basic fields in struct Client will
remain as they are.  However, instead of having a two-size struct Client,
I will split it into a struct Client and a struct Connection (or similar).

The first step to doing this is creating a set of accessor macros and
making all of ircd use them to access the fields.  This will allow me to
divorce the structures and only have to change the accessor macros and the
struct Client allocation functions and a bit of networking code, rather
than going through all of ircd later and fixing things after they're all
broken.

This commit is the first step toward this.  All the affected header files
have been converted, and all the .c files up through hash.c have been
carefully checked.  Everything compiles cleanly at this point, so I am
assuming that it works; I'll double-check when this phase of the project is
complete.


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/trunk@334 c9e4aea6-c8fd-4c43-8297-357d70d61c8c