Author: Kev <klmitch@mit.edu> Log message:
This finishes phase one of the struct Client/struct Connection divorce.
Everybody is now using the accessor macros.  IT IS VERY IMPORTANT THAT
ANY NEW CODE USE THESE MACROS!  Phase two will be much smaller, involving
only a new structure and changes to some accessor macros in client.h, and
some code moving from list.c to client.c with addition of struct Connection
allocation.

It somehow seems appropriate that MIT's mascot is the beaver.

Testing:

I've compiled and run the server and can't see any differences in behavior.
This is, however, a huge change, so something may very well be messed up.
Brute-force is the order of the day.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@336 c9e4aea6-c8fd-4c43-8297-357d70d61c8c