Author: Kev <klmitch@mit.edu> Log message:
This is it!  Connection information in struct Client has been pulled out
into its own structure, struct Connection.  I've also added a couple of
fields to struct Connection to enable me to create a more efficient
implementation of flush_connections() in the future...

Testing:

Compiles and runs without any apparent problems.  We need to bombard it
with clones up the wazoo to verify the integrity of make_client() and
free_client().


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@337 c9e4aea6-c8fd-4c43-8297-357d70d61c8c