Author: Diane Bruce <db@tfm.com> (by way of Kev <klmitch@mit.edu>) Log message:
- rfc1489 is ambigous about how to handle ISON. The example shows
  no ':' used in the command, which means each nick would appear in
  parv[1] to parv[n].  After discussion with hop on efnet/undernet
  it became apparent there is a problem. ircnet gets it right, but they
  loop over all parv as well as strtoken'ing on ' '
  My suggested fix is to loop over each parv[1] to parc, and scanning
  each one for ' '. I prefer to use parc rather than relying on a parv[]
  being NULL, but you could do a slightly more efficient version that way.

- Trailing spaces are removed, apparently that breaks some ircII clients.
  epic according to hoppie, is ok, but still...

(This patch was applied by hand because of the number of changes involved;
hope I got the trailing space behavior correct--I removed all trailing
spaces. -Kev)

Testing:

I tested it; it doesn't crash the server.  If ISON gets more than the
maximum number of arguments before hitting a colon, the server tries to
look up :<nick>, but this isn't much of a suprise.  Looks nice :)

 -Kev


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@338 c9e4aea6-c8fd-4c43-8297-357d70d61c8c