Author: Kev <klmitch@mit.edu> Log message:
Implement an ACL for IRC operators; set ACL out of information provided by
the .conf through the feature system.

Testing:

It's been tested on one server all by its lonesome, and seems to work fine;
it needs to be rigorously tested on a network with other users, and by both
local and global operators, with various fiddling of the /set's.  This big
a change to the privileges structure can lead to interesting anomolies.
Unfortunately, documentation is sparse; I'll look into correcting that when
I can, but don't hold your breath.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@344 c9e4aea6-c8fd-4c43-8297-357d70d61c8c