Author: Kev <klmitch@mit.edu> Log message:
Gee, I keep thinking I'm done with these files, and I'm not :)  This should
do the trick; most of the work was done in the name of future threading
capability--code it now rather than try to remember how I was going to do
it 3 months from now.  Clean up all compile-time warnings/errors, even in
the threading skeleton.

I'm pretty sure that I can now go to creating the event modules for kqueue,
/dev/poll, poll(), and select(), and then I'll probably resurrect
ircd_network.[ch] and put real networking code in it.  I intend to abstract
out the address information, so we should be able to deal, at least at this
layer, with IPv6 addresses, making it easier to deal with them in the
database later.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11_alpha_events@371 c9e4aea6-c8fd-4c43-8297-357d70d61c8c