Author: Kev <klmitch@mit.edu> Log message:
Based on discussion with Bleep, and the change in /wallops implementation
that has taken place on u2.10.10.pl14, sendcmdto_flag_butone() has been
recast to broadcast them to all servers, regardless of what sorts of users
those servers have on/behind them.  This reduces CPU wasted on /wallops to
the minimum practical, with only a small increase in protocol overhead--in
my opinion, that overhead is not enough to worry about.


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/trunk@415 c9e4aea6-c8fd-4c43-8297-357d70d61c8c