Author: Kev <klmitch@mit.edu> Log message:
Event-ize signals, try_connections(), and check_pings(); correct a small
bug in the routine that looks up the struct Signal for ET_SIGNAL
generation.  The next thing to be done before we start debugging all of
this is to convert the client routines in s_bsd.c to use the event engine
and remove the old read_message() code.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11_alpha_events@452 c9e4aea6-c8fd-4c43-8297-357d70d61c8c