Author: Kev <klmitch@mit.edu> Log message:
These are the last changes for the events system, sans debugging.  Some
of them are really hackish--I really dislike the timer I put in for running
queued client commands.  The next step will be fixing all the compile-time
errors and warnings--and I'm sure there are a lot of those.  Then comes the
debugging of the actual actions taken, along with the ever important
engines.  Finally, we can merge this all back into the trunk...and debug it
all over again :)


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11_alpha_events@453 c9e4aea6-c8fd-4c43-8297-357d70d61c8c