Author: Kev <klmitch@mit.edu> Log message:
Did some more debugging and found (and fixed) an off-by-one error in the
routine in engine_poll.c that adds a new socket to be considered.  Also
removed the once-a-second timer for client-throttling; we now do it on a
client-by-client basis, and only as needed.  Added /STATS e to report the
engine being used (so we can see when there's fallback behavior).  Added
Debug() calls all over the place.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11_alpha_events@462 c9e4aea6-c8fd-4c43-8297-357d70d61c8c