Author: Kev <klmitch@mit.edu> Log message:
As with any complicated piece of software, subtle bugs can creep in because
of simple things.  In this case, we have to keep track of when writes to a
client are blocked.  Which means that when we get an ET_WRITE event for
that client, the client isn't blocked anymore, so we have to turn off
FLAGS_BLOCKED.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11_alpha_events@473 c9e4aea6-c8fd-4c43-8297-357d70d61c8c