Author: Kev <klmitch@mit.edu> Log message:
When HEAD_IN_SAND_SNOTICES is #define'd, there is no longer any reason to
hide the nickname of an operator using, say, /OPMODE (except in the actual
MODE command sent to the channel).  Therefore, whenever
HEAD_IN_SAND_SNOTICES is #define'd, /OPMODE, /CLEARMODE, /GLINE, and /JUPE
generate server notices containing the generating operator's nickname.

The old chroot() and setuid() support code has been removed.  There are
plenty of code paths that get traversed before we even hit the setuid
stuff, which is a security risk.  There're much better ways of handling
setuid operation in the first place, which is what we're going to move to.
I will write documentation and preparation scripts for this.

Previously, whenever a user that is GLINEd or a server that is JUPEd showed
up on line from a remote server, the GLINE or JUPE was retransmitted.  To
prevent the network from being clogged up, I added a parameter to allow a
server to acknowledge that that JUPE/GLINE is known, and being
intentionally ignored.  After some discussion, however, it was decided to
remove all of that code.  Now the only retransmits left in the code are
ones where the remote server sends a JUPE/GLINE with a bad mod time...and
all that hair is now gone.


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/trunk@492 c9e4aea6-c8fd-4c43-8297-357d70d61c8c