Author: Kev <klmitch@mit.edu> Log message:
Implement a mechanism for servers to indicate to the network that they are
hubs, servers, or both.  We take parv[7], an unspecified unsigned int in
previous versions, and turn it into a mode-like string--it's not actually a
mode in the /mode #channel sense, but it looks similar.  This works in
conjunction with two new flags, FLAGS_HUB and FLAGS_SERVICE.  There is
currently no user-visible impact of this change.


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/trunk@493 c9e4aea6-c8fd-4c43-8297-357d70d61c8c