Author: Kev <klmitch@mit.edu> Log message:
I had a complaint that directed notices didn't work with u2.10.11.  Turns
out that the bug was in u2.10.10, which wasn't using numerics or tokens
for those notices!  This fixes pl15 to use tokens and numerics with
directed notices and mass-messages (foo@server.name.here and $*.org,
respectively).  All servers will have to upgrade to at least pl15 before
u2.10.11 can be deployed because of this :/


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_10_pl15@498 c9e4aea6-c8fd-4c43-8297-357d70d61c8c