Author: Kev <klmitch@mit.edu> Log message:
Adding my wrapper script, suitably re-commented and with the addition of
the ability to chroot, to the release.  This script allows one to set the
hard file descriptor limit (using -l), the root directory (using -c), and
the user ID (using -u), and runs a given command.  This program should NOT
be installed setuid, but must be run as root.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@512 c9e4aea6-c8fd-4c43-8297-357d70d61c8c