Author: Kev <klmitch@mit.edu> Log message:
This should be the last of the code required for u2.10.11 before release.
Not the last of the documentation, unfortunately, but you can't have
everything...Anyway, this implements restrictions on the G-line mask for
ordinary operators to keep accidental *@* G-lines from being set.  In
particular, if the host mask contains any wildcards or if the G-line
would impact 20 or more users (this number is configurable), then the
G-line would have to be forced by doing !+*@host.mask.goes.here or
whatever.  This check is also smart enough to not allow *@*foo.com or
*@127.0.0.1/15 bans, either.


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/trunk@521 c9e4aea6-c8fd-4c43-8297-357d70d61c8c