Author: Kev <klmitch@mit.edu> Log message:
Stomp some warnings discovered on NetBSD, on an alpha running NetBSD, and
under -O1 instead of -O3.  Remaining: "statement with no affect" in an
assert on line 489 of fda.c; pointer type mismatch warning for calls to
getsockopt() in the engines--due to NetBSD's use of socklen_t, which may
not always be available.  The former could be solved by removing the
assert and testing the condition during the configure stage; the latter
could be solved by defining socklen_t ourselves at configure time.


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/trunk@528 c9e4aea6-c8fd-4c43-8297-357d70d61c8c