Author: Kev <klmitch@mit.edu> Log message:
IPcheck sucks large, hairy rocks.  In this case, local clients were not
getting FLAGS_IPCHECK set, so the connected count in the IP registry was
not getting decremented when they exited.  Moreover, IPcheck_remote_connect
set this bit itself, but IPcheck_local_connect cannot because the client
doesn't even exist yet.  This means that the bit has to be set in s_bsd.c.
This is pretty darned annoying, but what else can one do?  This should at
least fix one problem we've been having with the IPcheck code.


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/trunk@530 c9e4aea6-c8fd-4c43-8297-357d70d61c8c