Author: Kev <klmitch@mit.edu> Log message:
If servers are disconnected through a call to exit_client()--say by a squit
or something else, perhaps an access check--the server would send a raw
message like "ERROR :Closing Link [...]".  The server interprets "ERROR" as
a numeric nick, and since in most cases, that numeric nick doesn't exist,
it ignores the message.  This means that servers that closed the connection
frequently did so without the proper (and useful) error message.  This fix
simply uses sendrawto_one() only for clients, and uses sendcmdto_one() for
servers, fixing the error.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@534 c9e4aea6-c8fd-4c43-8297-357d70d61c8c