Author: mbuna <mbuna@undernet.org> Log message:
Darwin is the core of Mac OS X. It's a real BSD, and its very similar
to freeBSD on many points. Actually, ircu's config systems doesnt
recognize it. Users can force with the --host option, but the build
will fail due to minor compatibility problems on os_generic. It's
working well when linking with os_bsd.c.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@561 c9e4aea6-c8fd-4c43-8297-357d70d61c8c