Author: Kev <klmitch@mit.edu> Log message:
Fix the SQUIT bug.  Originally, we were treating the server name argument
as a numeric nick, and if we couldn't find a server that way, then we'd try
looking it up as a bare server name.  Unfortunately, FindNServer() doesn't
check lengths; pass it "Amsterdam2.NL.EU.undernet.org" and it gives you the
server with numeric "Am"--stockholm.se.eu.undernet.org.  The reason we were
getting a timestamp mismatch was because we were checking the wrong server.
The fix is simply to reverse the order we use to decode the server name
argument--try it as a bare name first (FindServer), *then* if that isn't
successful, try decoding it as a numeric nick.  Numeric nicks will never be
valid server names, so we shouldn't have problems with this.  This hasn't
even been compile-tested, but the fix is intuitively obvious.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11@593 c9e4aea6-c8fd-4c43-8297-357d70d61c8c