Author: Kev <klmitch@mit.edu> Log message:
Implementation of ACCOUNT.  This simple feature associates an account name
with a user and provides means for setting and reading this account name.
It is intended for use with the various service bots on Undernet, and
allows for seamless re-login when the net reconnects after a split.  Here's
how it works:

When a user logs in to, say, X, the X server will issue an ACCOUNT command,
which will be broadcast across the network, to set the client's account
name to the username he logged in to.  From now on, whenever the client is
referenced by a NICK message, the special usermode '+r' will be set.  This
usermode takes an argument, the account name as set by ACCOUNT.  (The mode
may not be set or unset by the user for security reasons.)  This means that
when X reconnects to the network, it is told what username the user is
logged in to, so the user no longer has to re-login.  This also means that
the user is also automatically logged in to other service bots on the
network.  If we load a backup channel service, the user will be able to log
in to both by sending a single login command to one bot.

The account name shows up in /whois with numeric 330.  The parameters are,
beyond the basic numeric reply parameters: queried user's nickname (as for
all /whois replies), the account name, and the static string "is logged in
as".  This reply is only generated if the user's account name has been set.

Testing status:

The code compiles cleanly; I have verified that sending ACCOUNT properly
sets the account name for the user, that /whois reports the correct account
name, that NICKs generated during the burst have the proper +r mode with
its argument, and that NICKs with the +r mode followed by the account name
argument properly set the account name.  I've also verified that bogus
invocations of ACCOUNT result in the correct error messages.  I have not
verified proper propagation of ACCOUNT, nor have I verified that bounds-
checking works properly, but I'm pretty certain that both will work
correctly.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11@594 c9e4aea6-c8fd-4c43-8297-357d70d61c8c