Author: Kev <klmitch@mit.edu> Log message:
Start to try to track down problems seen when running u2.10.11 on
Amsterdam2 that look like some sort of list corruption problem.  This
patch is only effective if --enable-debug and --enable-asserts are
specified, and will (unfortunately) be something of a CPU hog.  This is
essentially the sledgehammer approach to finding the problem :)


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/branches/u2_10_11@600 c9e4aea6-c8fd-4c43-8297-357d70d61c8c