Author: Kev <klmitch@mit.edu> Log message:
Remove a bogus assert; only call release_auth_client() from
destroy_auth_request() if send_reports is 1.  This can be done because
destroy_auth_request() is only called in two places--during a time-out,
where reports should be sent, and in the call to free_client(), where
reports should not be sent, and where the client should not be added to
the client linked list.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11@602 c9e4aea6-c8fd-4c43-8297-357d70d61c8c