Author: Kev <klmitch@mit.edu> Log message:
Looked for all the hiding changes between pl15 and pl18 and forward-ported
those that were not already in u2.10.11.  Compiled and running on testnet.
Looks good, and I believe I got it all.  I'm starting to feel that beta.07
is a fully-qualified release candidate...once it proves stable on the
network, of course.


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/branches/u2_10_11@606 c9e4aea6-c8fd-4c43-8297-357d70d61c8c