Author: Kev <klmitch@mit.edu> Log message:
I think I finally have it.  The bug was caused by the fact that I was
using a next pointer.  The timer that next in timer_run() was pointing at
could just plain go away.  But there's no need to have used a next pointer;
the algorithm ensures that that next timer to process is always at the head
of the queue!


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/branches/u2_10_11@621 c9e4aea6-c8fd-4c43-8297-357d70d61c8c