Author: Carlo Wood <run@alinoe.com> Log message:
Implementation of network synchronization code.                                 Includes the implementation of the previously                                   unused DESTRUCT message.  Correction on ms_join                                 to copy older timestamps and now allowing                                       BURST message to occur outside the net-burst.                                   See my last two posts to coder-com for detailed                                 information on how I got to this implementation                                 and for a C++ simulation program that wrote to                                  test it.                                                                                                                                                        Note that I changed the copyright notice in                                     m_destruct - Jarko didn't add that function, I did.                             Not that it is a big deal to me.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@668 c9e4aea6-c8fd-4c43-8297-357d70d61c8c