Author: Kev <klmitch@mit.edu> Log message:
If a user changes nickname to a nick that's banned on the channel, the ban
wouldn't take effect.  Although minor, this is a bug, and may best be fixed
by invalidating all the ban caching for the user.  This is effected by
looping over all the user's channel memberships and doing a ClearBanValid()
on them.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11@684 c9e4aea6-c8fd-4c43-8297-357d70d61c8c