Author: Kev <klmitch@mit.edu> Log message:
What a screw case.  In the case of many, many joins, we were overflowing an
array in the joinbuf because we were adding a channel to the array, then
forgetting to empty the array when we flushed--because they weren't
supposed to have been added in the first place!  This should stop them from
being added anymore, and hopefully fix our Amsterdam* crashes.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11@695 c9e4aea6-c8fd-4c43-8297-357d70d61c8c