Author: Stephane Thiell <mbuna@undernet.org> Log message:
Sorry for the previous patch, it didn't work. This one definitely fix
the problem on remote whois. The bug was amazing: parv[1] is always a
server numeric, and with FindClient, we were seeking nickname.
Conclusion: if a nick on the network was a numnick of a server, the
remote whois was almost always badly relayed.


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/branches/u2_10_11@725 c9e4aea6-c8fd-4c43-8297-357d70d61c8c