Author: Kev <klmitch@mit.edu> Log message:
I got tired of the abortion known as /STATS, so I re-implemented it in a
much cleaner fashion.  Instead of using a massive switch, each /STATS gets
its own function, which is listed in a dispatch table with various flags
available.  The dispatch table also includes documentation on each of the
/STATS letters it recognizes.  A stats_init() function initializes an array
indexed on the /STATS letters so that we don't have to do a linear search
of the array to find the entry we want.  This patch includes support for
the up-coming policy featurization, but this is hacked out temporarily so
that it'll compile.

Benefits:
 * You can finally understand what the heck m_stats() is doing!
 * Massive reduction of code duplication.
 * Centralization of access control.

Testing:

It compiles and works correctly.  Most of the code is not in fact new, and
the new code is very short and simple, so there were not a whole lot of
bugs that cropped up during development--most of those that did were simple
compile-time errors or warnings that were easy to correct.  Note that this
does *not* obey ircd_policy.h at the moment!


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11@794 c9e4aea6-c8fd-4c43-8297-357d70d61c8c