Author: Kev <klmitch@mit.edu> Log message:
The new /stats code would not propagate a /stats flag it did not recognize
to a remote server.  I considered this a bug, since that's one of the
essential features of /stats.  This fixes this by adding a /stats flag '*'
that sends the help information to the (local) user.  If we cannot find the
StatDesc structure for the letter given by the user, we substitute the one
for '*' instead.  The side-effect of this is that any /stats flag that the
server doesn't understand falls into the '*' bucket and gets propagated by
the same code that propagates the understood /stats flags.  Also, the
descriptor for '*' has the STAT_FLAG_VARPARAM flag set, allowing us to
propagate a single extra parameter for said flag, which the old /stats code
couldn't do.


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/branches/u2_10_11@801 c9e4aea6-c8fd-4c43-8297-357d70d61c8c