Author: Kev <klmitch@mit.edu> Log message:
Feature names in ircd.conf used to be case insensitive.  Unfortunately,
the HIS_STATS_? features require that the feature names be case sensitive
in order for them to be configured properly.  If you are using lower-case
or mixed-case feature names in your ircd.conf, THIS WILL BREAK THEM.
Sorry, but this is the easiest way to fix the problem...


git-svn-id: https://undernet-ircu.svn.sourceforge.net/svnroot/undernet-ircu/ircu2/branches/u2_10_11@807 c9e4aea6-c8fd-4c43-8297-357d70d61c8c