Author: Kev <klmitch@mit.edu> Log message:
First step in deprecation of modeless channels--don't permit their
creation.  You can still join an existing modeless channel, but since no
one actually uses them, they'll die a quick death as soon as the net is up
to .11--and then we can get rid of the rest of the code.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11@808 c9e4aea6-c8fd-4c43-8297-357d70d61c8c