Author: Kev <klmitch@mit.edu> Log message:
Most of these changes are to support a kinder, gentler timestamp drift
correction algorithm in ms_create().  The basic idea is that if the
timestamp drift is bad enough for the server to send a server notice about
it, then it *also* sends a SETTIME to correct the time.  Now this alone is
insufficient if we're the server that's off--that's where the modifications
to m_settime.c come in.  We now allow servers to send SETTIMEs to specific
remote servers, rather than only spamming the network.  If we handle such
a directed SETTIME that is not directed to *us*, and if we're
RELIABLE_CLOCK and the timestamp on the incoming SETTIME was more than 30
seconds out of bounds, then we send a corrective SETTIME back up the line
to our previous hop.  This requires all of the hubs to be RELIABLE_CLOCK in
order to work properly.  The mod to s_user.c is necessary to remove an
assertion that prevented hunt_server_prio_cmd() from being used with
messages originating from a server.

A side-effect of this change, since I rewrote mo_settime() from scratch, is
to allow opers to specify a timestamp of 0 to SETTIME--if the oper-
specified timestamp is 0 or if we're a RELIABLE_CLOCK server, then our
current time is substituted instead.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11_03@866 c9e4aea6-c8fd-4c43-8297-357d70d61c8c