Author: Kev <klmitch@mit.edu> Log message:
Finally, the long awaited database primitives library is committed.  This
is missing some test cases for the sparse matrices--I'm not going to have
much time in the near future, so I figured I should at least try to solicit
people's help with that aspect of it.  The library also includes complete
(hopefully) documentation, generated by doxygen--see the doc subtree for
that.  Works well, as far as the test cases I've written are concerned :)
The auto-resizing hash tables will be among the biggest wins for ircu, as
will the sparse matrices.


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@923 c9e4aea6-c8fd-4c43-8297-357d70d61c8c