Author: hikari <shadow@undernet.org> Log message:
	Hopefuly fixed the bug whereby users couldn't connect to Amsterdam2*
after a certain amount of time - in a eureka momment realsised this was
probably due to the larger number of clients distributing the check_pings()
interval a little more tightly (hence why we couldn't replicate it).

	Non registered clients are now checked against their firsttime()
(ie when they connected) and the CONNECTTIMEOUT value for the server and
bumped off with a message stating "Registration Timeout".  The code to
actually disconnect the unregistered clients was there, it just wasn't
being utilised correctly.

	Hopefuly this will be the last pre-release before the actual
release.

blessed be,
hikari


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_11_05@955 c9e4aea6-c8fd-4c43-8297-357d70d61c8c