Author: Kev <klmitch@mit.edu> Log message:
Implement a custom assert() macro (in ircd_log.h).  This variant of
assert() is similar to the system assert(), in that it ends with a call to
abort() if the assertion fails; however, logging is done through the
logging subsystem.  (A sentinel is added to (hopefully) prevent an
assertion failure somewhere in the logging subsystem from entering into an
infinite loop; if this happens, there will be no output at all, but abort()
will still be called.)


git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@1271 c9e4aea6-c8fd-4c43-8297-357d70d61c8c